usePlayer: extract continue-playing settings mirror into its own hook
The "continue playing after queue ends" backend setting was
mirrored as a live-updating ref via a 30-line useEffect that
fetched on mount AND listened for `tidal-settings-updated` window
events from SettingsPage. Self-contained — only used by the
`continuePlayingRef.current` reads in advanceRef and applySnapshot's
error-recovery branch.

Moved to `useContinuePlayingPref()` which returns the ref directly.
usePlayer is one line: `const continuePlayingRef =
useContinuePlayingPref()`.

Future generalisation note: this is the same shape as any other
"backend setting mirrored as a live ref so the SSE / advance path
can read it sync without re-installing on every settings change".
If a second one shows up we can promote to a generic
`useSettingsField<K>(key)`. For now, one specific hook is simpler.

usePlayer.ts: 1077 → 1041 LOC. tsc clean. vitest 36 passed.
